permuter: improve MemHeap::Print (89.7% -> 90.0%)
signunsign cast on unsigned hdr zero-check in Print loop.

+++ src/system/utl/MemHeap.cpp
@@ -92,7 +92,7 @@ void MemHeap::Print(TextStream &ts, bool verbose) {
             // Alloc block
             unsigned int hdr = *curPtr;
             unsigned int *headerPtr = curPtr;
-            while (hdr == 0) {
+            while ((int)hdr == 0) {
                 headerPtr++;
                 hdr = *headerPtr;
             }
